rewrite of the psm02 driver
this pretty much reverts Commit cde6ce4e150e72c56fb4cb7f05778a5064b36960
from Feb 19 2017 for this driver which breaks it. (also version 0.0.14)
(he's been busy with this for a long time it seems like)

report parsers did not check what sensor type was coming in, now they do

report parsers of the contact and tamper alarms were from alarm sensor
command class, which is not supported, but also not used in the
"command_class" key.
changed it to the appropiate report parser.

removed get parsers in the temperature and luminance capabilities, they
are not used without a command_get, but also not very applicable for
battery devices

removed the battery get on awake, it is already reported every 6 hours
on default, changeable in the settings.

made the code more JS-Lint compliant

added tamper cancelation (this is not done by default), copy paste the
(tested) code from psp05